// takin-webapp/tro-web-ui/src/common/menu/type.ts
/*
 * Copyright 2021 Shulie Technology, Co.Ltd
 * Email: shulie@shulie.io
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * @name 菜单项
 */
export interface MenuBean {
  model?: string;
  title: string;
  icon?: string;
  path?: string;
  key?: string;
  type?: MenuType;
  authority?: any;
  hideInMenu?: boolean;
  children?: MenuBean[];
}
/**
 * @name 菜单类型
 */
export enum MenuType {
  SubMenu = 'SubMenu',
  ItemGroup = 'ItemGroup',
  Item = 'Item',
  Url = 'url',
  NoMenu = 'NoMenu'
}